refactor(providers): extract shared provider auth helpers module (#180)
